requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
flask_cors
bokeh
selenium
plotly
numpy
scipy
pandas
sqlalchemy
python-dotenv
pymysql 